fixing RedisCloudDatabases
